Moved example under chat directory, refactored code as per comments
