+++ unity/Assets/Ably/Examples/Chat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 52c70a000ba56504792c888679ab3ebc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
