+++ unity/Assets/Ably/Examples/Chat/AblyChannelUiConsole.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6f24b1992fd53cd4f8111263f1add2e0
+guid: 9736114b038d2ad4584f69f64a5d7006
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
