+++ unity/Assets/Ably/Examples/Chat/AblyConsole.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a7cc2f71ec7840843897449abc1c9120
+guid: 29f9185da29c4b34f92c086a254d3110
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
