+++ unity/Assets/Ably/Examples/Chat/AblyPresenceUiConsole.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a6e2024162482c84a96750f78d375aa5
+guid: 8735a2c1cf276864891d2495215f1f7e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
