+++ unity/Assets/Ably/Examples/Chat/logo.png
